fixes CB-143 removing address from iOS contact causes crash
Problem was in setMultiValueDictionary: The address multi value was being removed from within the for loop that updates the individual fields rather than after the for loop completed.  Also fixed a problem with address type - this was being incorrectly changed to "other" when the input value was an empty string with the end result that the address was not removed when it should have been.